tg_bot/requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
aiogram==3.15.0
langchain-gigachat==0.3.2
langchain-community==0.3.13
tiktoken==0.5.2